requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
streamlit
torch==1.8.1+cpu 
torchvision==0.9.1+cpu
fastai>=2.3.1
ipywidgets
voila
pillow==8.2.0